--- src/client/src/components/Button/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Button() {
-  return (
-    <div>index</div>
-  )
-}
-
-export default Button
