--- src/client/src/components/GlobalStyles/index.js
@@ -1,7 +1,0 @@
-import './GlobalStyles.css'
-
-function GlobalStyles({ children }) {
-  return children;
-}
-
-export default GlobalStyles;
