--- src/client/src/layouts/DefaultLayout.jsx
@@ -1,15 +1,0 @@
-import { Outlet } from "react-router-dom"
-import Sidebar from '../components/Sidebar'
-
-function DefaultLayout() {
-  return (
-    <div>
-      <Sidebar />
-      <div className="content">
-        <Outlet />
-      </div>
-    </div>
-  )
-}
-
-export default DefaultLayout
