--- src/client/src/pages/TaskManagement/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function TaskManagement() {
-  return (
-    <div>TaskManagement</div>
-  )
-}
-
-export default TaskManagement
