+++ src/client/src/pages/homepage/index.js
@@ -1,0 +1,11 @@
+import Sidebar from '../../components/sidebar'
+
+function HomePage() {
+  return (
+    <div>
+      <Sidebar />
+    </div>
+  )
+}
+
+export default HomePage
